adding pread64 simprocedure
