+++ angr/procedures/definitions/linux_kernel.py
@@ -12,6 +12,7 @@ lib.add('close', P['posix']['close'])
 lib.add('exit', P['libc']['exit'])
 lib.add('sigaction', P['posix']['sigaction'])
 lib.add('dup', P['posix']['dup'])
+lib.add('pread64', P['posix']['pread64'])
 
 lib.add_alias('exit', 'exit_group')
 lib.add_alias('getuid', 'geteuid')
